counsel.el: Add counsel-info-apropos, counsel-info-apropos-for-manual.
These functions augment info-apropos to with an ivy interface.
counsel-info-apropos will search just like info-apropos but with a
dynamic-collection that searches again each time the input is changed.
counsel-info-apropos-for collects all entries for a manual that can be
narrowed with ivy.  Both use a timer and generator to keep the
minibuffer input fast and responsive.

* counsel.el (counsel-info-apropos--manuals): New function to get all
known manuals.
* counsel.el (counsel-info-manual--buffer-name): New function to
compute the temp buffer name used to search an Info manual.
* counsel.el (counsel-info-manual--matches): New generator to search
all entries in a manual.
* counsel.el (counsel-info-apropos--format-candidate): New function to
format candidates of counsel-info-apropos and
counsel-info-apropos-for-manual.
* counsel.el (counsel-info-apropos--node-match-p): New predicate to
check if a candidate Info entry matches the search criteria.
* counsel.el (counsel-info-apropos-timer): New variable that stores
the timer used to collect Info entries.
* counsel.el (counsel-info-apropos-state): New struct type to hold
buffer and iterator for Info searches.
* counsel.el (counsel-info-apropos--set-candidates): New function to
set ivy-candidates and update counsel-info-apropos-state.
* counsel.el (counsel-info-apropos--start-timer): New function to
start a timer that collects Info search results.
* counsel.el (counsel-info-manual-apropos--handle-input): New function
to handle input for counsel-info-manual-apropos.
* counsel.el (counsel-info-apropos--goto-selection): New function to
jump to a search result as the :action of counsel-info-apropos and
counsel-info-manual-apropos.
* counsel.el (counsel-info-apropos--cleanup-buffer-for-manual): New
function to cleanup counsel-info-apropos search buffers.
* counsel.el (counsel-info-apropos--cleanup-timer): New function to
cleanup timer for collecting search results.
* counsel.el (counsel-info-apropos-for-manual): New command to search
all entries in a manual.
* counsel.el (counsel-info-manual-apropos): New command to select an
Info manual and then search all entries in the manual.
* counsel.el (counsel-info-apropos--matches): New iterator to search
all Info manuals for a pattern.
* counsel.el (counsel-info-apropos): New command to add an ivy
interface to info-apropos.